Only save tmux server if server is open

+++ config/i3_ubuntu/logout.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
 # Save tmux session and exit server
-~/.tmux/plugins/tmux-resurrect/scripts/save.sh
-sleep 1
-tmux kill-server
+if tmux info &> /dev/null; then
+  ~/.tmux/plugins/tmux-resurrect/scripts/save.sh
+  tmux kill-server
+fi
 
 # Exit i3 session
 i3-msg exit
